+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Bug #17: Add references to predefined enum values in `HasStrokeLineCap` and `HasStrokeLineJoin` traits (@terabytesoftw)
 - Bug #18: Update assertion message for integer setting in `StrokeWidthProvider` class (@terabytesoftw)
 - Enh #19: Add `HasStrokeDashArray` trait and corresponding tests for managing SVG `stroke-dasharray` attribute (@terabytesoftw)
+- Bug #20: Update `HasStrokeDashArray` trait and tests to support float values for `stroke-dasharray` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024